Add files via upload
Abstract:
This project aims to identify the features of a given and old or new South African banknote (R10, R20, R50, R100, R200) referencing a collection of notes. It considers the sides (back and front), scale, and rotation of the note to identify a similarity between the given note and reference notes. This is done by preprocessing our given bank notes dataset, enhancing, segmenting, performing feature extraction, and classifying the banknote. The proposed system will be able to work on old and new South African banknotes. Since the South African banknotes are easily classifiable based on the color and the texture, hence then we propose a system that will use GLCM to extract the texture of the banknote, and the colour of the image will be selected based on the RGB spaces. When it comes to classifying each note based on the denomination, it will be using SVM and KNN. The performance of these classifiers was evaluated in terms of training performance and classification accuracies.